refactor: change folders inside domain folder to improve DDD organization based on bounded contexts
