+++ src/domain/customer/event/handler/send-email-2-when-customer-is-created.handler.ts
@@ -1,0 +1,9 @@
+import EventHandlerInterface from "@domain/@shared/event/event-handler.interface";
+import CustomerCreatedEvent from "@domain/customer/event/customer-created.event";
+
+export default class SendEmail2WhenCustomerIsCreatedHandler
+  implements EventHandlerInterface<CustomerCreatedEvent> {
+  handle(event: CustomerCreatedEvent): void {
+    console.log("Esse é o segundo console.log do evento: CustomerCreated");
+  }
+}
