+++ src/domain/customer/repository/customer-repository.interface.ts
@@ -1,0 +1,5 @@
+import RepositoryInterface from "../../@shared/repository/repository-interface";
+import Customer from "@domain/customer/entity/customer";
+
+export default interface ICustomerRepository
+  extends RepositoryInterface<Customer> {}
