--- src/domain/event/@shared/event-dispatcher.spec.ts
@@ -1,65 +1,0 @@
-import EventDispatcher from "./event-dispatcher";
-import SendEmailWhenProductIsCreatedHandler from "@domain/event/product/handler/send-email-when-product-is-created.handler";
-import ProductCreatedEvent from "@domain/event/product/product-created.event";
-
-describe("Domain events tests", () => {
-  it("should register an event handler", () => {
-    const eventDispatcher = new EventDispatcher();
-    const eventHandler = new SendEmailWhenProductIsCreatedHandler();
-
-    eventDispatcher.register("ProductCreatedEvent", eventHandler);
-
-    expect(eventDispatcher.eventHandlers["ProductCreatedEvent"]).toBeDefined();
-    expect(eventDispatcher.eventHandlers["ProductCreatedEvent"].length).toBe(1);
-    expect(eventDispatcher.eventHandlers["ProductCreatedEvent"][0]).toMatchObject(eventHandler);
-  });
-
-  it("should unregister an event handler", () => {
-    const eventDispatcher = new EventDispatcher();
-    const eventHandler = new SendEmailWhenProductIsCreatedHandler();
-
-    eventDispatcher.register("ProductCreatedEvent", eventHandler);
-    expect(eventDispatcher.eventHandlers["ProductCreatedEvent"]).toBeDefined();
-
-    eventDispatcher.unregister("ProductCreatedEvent", eventHandler);
-    expect(eventDispatcher.eventHandlers["ProductCreatedEvent"]).toBeDefined();
-    expect(eventDispatcher.eventHandlers["ProductCreatedEvent"].length).toBe(0);
-  });
-
-  it("should unregister all event handlers", () => {
-    const eventDispatcher = new EventDispatcher();
-    const eventHandler = new SendEmailWhenProductIsCreatedHandler();
-
-    eventDispatcher.register("ProductCreatedEvent", eventHandler);
-    expect(eventDispatcher.eventHandlers["ProductCreatedEvent"]).toMatchObject(eventHandler);
-
-    eventDispatcher.unRegisterAll();
-    expect(eventDispatcher.eventHandlers["ProductCreatedEvent"]).toBeUndefined();
-  });
-
-  it("should notify an event", () => {
-    const eventDispatcher = new EventDispatcher();
-    const eventHandler = new SendEmailWhenProductIsCreatedHandler();
-    const spyEventHandler = jest.spyOn(eventHandler, "handle");
-
-    eventDispatcher.register("ProductCreatedEvent", eventHandler);
-
-    expect(
-      eventDispatcher.eventHandlers["ProductCreatedEvent"][0]
-    ).toMatchObject(eventHandler);
-
-    const productCreatedEvent = new ProductCreatedEvent({
-      id: 1,
-      name: "Product 1",
-      price: 10,
-      quantity: 10,
-    });
-
-    // When notify is called, the
-    // SendEmailWhenProductIsCreatedHandler.handle() should be called
-    eventDispatcher.notify(productCreatedEvent);
-
-    expect(spyEventHandler).toHaveBeenCalled();
-
-  });
-});
