--- src/domain/event/customer/customer-created.event.ts
@@ -1,10 +1,0 @@
-import EventInterface from "@domain/event/@shared/event.interface";
-
-export default class CustomerCreatedEvent implements EventInterface {
-  dataTimeOccured: Date;
-  eventData: any;
-  constructor(data: any) {
-    this.dataTimeOccured = new Date();
-    this.eventData = data;
-  }
-}
