--- src/domain/event/product/product-created.event.ts
@@ -1,10 +1,0 @@
-import EventInterface from "@domain/event/@shared/event.interface";
-
-export default class ProductCreatedEvent implements EventInterface {
-  dataTimeOccured: Date;
-  eventData: any;
-  constructor(data: any) {
-    this.dataTimeOccured = new Date();
-    this.eventData = data;
-  }
-}
