+++ src/domain/product/event/handler/send-email-when-product-is-created.handler.ts
@@ -1,0 +1,9 @@
+import EventHandlerInterface from "@domain/@shared/event/event-handler.interface";
+import ProductCreatedEvent from "@domain/product/event/product-created.event";
+
+export default class SendEmailWhenProductIsCreatedHandler
+  implements EventHandlerInterface<ProductCreatedEvent> {
+  handle(event: ProductCreatedEvent): void {
+    console.log("Sending email to ......");
+  }
+}
