+++ src/domain/product/repository/product-repository.interface.ts
@@ -1,0 +1,5 @@
+import RepositoryInterface from "../../@shared/repository/repository-interface";
+import Product from "@domain/product/entity/product";
+
+export default interface IProductRepository
+  extends RepositoryInterface<Product> {}
