--- src/domain/repository/customer-repository.ts
@@ -1,5 +1,0 @@
-import RepositoryInterface from "./repository-interface";
-import Customer from "@domain/entity/customer";
-
-export default interface ICustomerRepository
-  extends RepositoryInterface<Customer> {}
