--- src/domain/repository/product-repository.ts
@@ -1,5 +1,0 @@
-import RepositoryInterface from "./repository-interface";
-import Product from "@domain/entity/product";
-
-export default interface IProductRepository
-  extends RepositoryInterface<Product> {}
