+++ src/main.ts
@@ -1,8 +1,8 @@
 import 'module-alias/register';
-import Customer from "@domain/entity/customer";
-import Address from "@domain/entity/address";
-import OrderItem from "@domain/entity/order_item";
-import Order from "@domain/entity/order";
+import Customer from "@domain/customer/entity/customer";
+import Address from "@domain/customer/value-object/address";
+import OrderItem from "@domain/checkout/entity/order_item";
+import Order from "@domain/checkout/entity/order";
 
 function main () {
   const customer = new Customer("1", "John Doe");
